NatsFlow/requirements.txt
nats-py
python-dotenv
fastapi
uvicorn
